+++ examples/sass-plugin-rust/src/main.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export function Main() {
+  return (
+    <>
+      <div>
+        <div className="dep">sass</div>
+      </div>
+    </>
+  );
+}