+++ js-plugins/postcss/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @farmfe/js-plugin-postcss
 
+## 1.5.3
+
+### Patch Changes
+
+- f6196f14: fix postcss internal plugins
+- Updated dependencies [6425c763]
+  - @farmfe/core@1.0.9
+
+## 1.5.2
+
+### Patch Changes
+
+- cb7df71f: Support alias resolve and url rebase for import.meta.glob, sass and less plugins
+- Updated dependencies [cb7df71f]
+  - @farmfe/core@1.0.8
+
 ## 1.5.1
 
 ### Patch Changes